requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_iam
aws-cdk.aws_ssm
aws-cdk.aws_autoscaling
aws-cdk.aws_secretsmanager
aws-cdk.aws_config
aws-cdk.aws_lambda




